requirements.txt
duckframework
anthropic
google-generativeai
groq
